+++ client/network/sync/Cargo.toml
@@ -20,11 +20,16 @@ prost-build = "0.9"
 codec = { package = "parity-scale-codec", version = "3.0.0", features = [
     "derive",
 ] }
+futures = "0.3.21"
 libp2p = "0.44.0"
 log = "0.4.16"
 prost = "0.9"
 sc-client-api = { version = "4.0.0-dev", path = "../../api" }
 sc-consensus = { version = "0.10.0-dev", path = "../../consensus/common" }
+sc-network-common = { version = "0.10.0-dev", path = "../common" }
 smallvec = "1.8.0"
+sp-blockchain = { version = "4.0.0-dev", path = "../../../primitives/blockchain" }
 sp-core = { version = "6.0.0", path = "../../../primitives/core" }
+sp-finality-grandpa = { version = "4.0.0-dev", path = "../../../primitives/finality-grandpa" }
 sp-runtime = { version = "6.0.0", path = "../../../primitives/runtime" }
+thiserror = "1.0"
