+++ client/network/sync/src/lib.rs
@@ -20,3 +20,5 @@ pub mod blocks;
 pub mod message;
 pub mod schema;
 pub mod state;
+pub mod warp;
+pub mod warp_request_handler;
